Converting victory-scatter to TypeScript
