+++ packages/victory-scatter/src/helper-methods.tsx
