--- packages/victory-scatter/src/index.js
@@ -1,1 +1,0 @@
-export { default as VictoryScatter } from "./victory-scatter";
