+++ packages/victory-scatter/src/index.ts
@@ -1,0 +1,1 @@
+export * from "./victory-scatter";
