+++ packages/victory-scatter/src/victory-scatter.tsx
@@ -13,6 +13,14 @@ import {
   Domain,
   Point,
   UserProps,
+  EventPropTypeInterface,
+  ScatterSymbolType,
+  StringOrNumberOrCallback,
+  EventsMixinClass,
+  VictoryCommonProps,
+  VictoryDatableProps,
+  VictoryMultiLabelableProps,
+  VictoryStyleInterface,
 } from "victory-core";
 import { getBaseProps } from "./helper-methods";
 
@@ -24,7 +32,34 @@ const fallbackProps = {
   symbol: "circle",
 };
 
-class VictoryScatter extends React.Component {
+export type VictoryScatterTTargetType = "data" | "labels" | "parent";
+
+export interface VictoryScatterProps
+  extends VictoryCommonProps,
+    VictoryDatableProps,
+    VictoryMultiLabelableProps {
+  bubbleProperty?: string;
+  children?: React.ReactElement | React.ReactElement[];
+  events?: EventPropTypeInterface<
+    VictoryScatterTTargetType,
+    StringOrNumberOrCallback
+  >[];
+  eventKey?: StringOrNumberOrCallback;
+  maxBubbleSize?: number;
+  minBubbleSize?: number;
+  size?: number | { (data: any): number };
+  style?: VictoryStyleInterface;
+  symbol?: ScatterSymbolType | { (data: any): ScatterSymbolType };
+}
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+interface VictoryScatterBase extends EventsMixinClass<VictoryScatterProps> {}
+
+/**
+ * Draw area charts with React. VictoryArea is a composable component, so it doesn't include axes.
+ * Add VictoryArea as a child of VictoryChart for a complete chart.
+ */
+class VictoryScatterBase extends React.Component<VictoryScatterProps> {
   static animationWhitelist = [
     "data",
     "domain",
@@ -108,4 +143,4 @@ class VictoryScatter extends React.Component {
   }
 }
 
-export default addEvents(VictoryScatter);
+export const VictoryScatter = addEvents(VictoryScatterBase);
